Support SOL_Payload_Access configuration for each user in bmc-config

+++ libfreeipmi/src/ipmi-utils.c
@@ -117,14 +117,14 @@ ipmi_check_completion_code(fiid_obj_t obj_cmd, uint8_t completion_code)
       char errstr[ERR_WRAPPER_STR_MAX_LEN];
       snprintf (errstr, ERR_WRAPPER_STR_MAX_LEN,
                 "ipmi_check_completion_code: completion code invalid: %x",
-                completion_code_recv);
+                (uint8_t)completion_code_recv);
       syslog (LOG_MAKEPRI (LOG_FAC (LOG_LOCAL1), LOG_ERR), errstr);
     }
 #endif /* IPMI_SYSLOG */
 #if defined (IPMI_TRACE)
       fprintf(stderr,
               "ipmi_check_completion_code: completion code invalid: %x",
-              completion_code_recv);
+              (uint8_t)completion_code_recv);
       fflush(stderr);
 #endif /* IPMI_TRACE */
     }
